[actix_migration] Consolidate actix impl into single directory (#13986)
No functional change.
