+++ core/async/src/actix/mod.rs
@@ -1,0 +1,5 @@
+pub mod futures;
+mod sender;
+pub mod wrapper;
+
+pub use sender::*;
