+++ core/async/src/futures.rs
@@ -1,8 +1,6 @@
-use actix::Actor;
 use futures::FutureExt;
 pub use futures::future::BoxFuture; // pub for macros
 use near_time::Duration;
-use std::ops::DerefMut;
 use std::sync::Arc;
 
 /// Abstraction for something that can drive futures.
@@ -60,15 +58,6 @@ pub fn respawn_for_parallelism<T: Send + 'static>(
     async move { receiver.await.unwrap() }
 }
 
-/// A FutureSpawner that hands over the future to Actix.
-pub struct ActixFutureSpawner;
-
-impl FutureSpawner for ActixFutureSpawner {
-    fn spawn_boxed(&self, description: &'static str, f: BoxFuture<'static, ()>) {
-        near_performance_metrics::actix::spawn(description, f);
-    }
-}
-
 /// A FutureSpawner that gives futures to a tokio Runtime, possibly supporting
 /// multiple threads.
 pub struct TokioRuntimeFutureSpawner(pub Arc<tokio::runtime::Runtime>);
@@ -76,19 +65,6 @@ pub struct TokioRuntimeFutureSpawner(pub Arc<tokio::runtime::Runtime>);
 impl FutureSpawner for TokioRuntimeFutureSpawner {
     fn spawn_boxed(&self, _description: &'static str, f: BoxFuture<'static, ()>) {
         self.0.spawn(f);
-    }
-}
-
-pub struct ActixArbiterHandleFutureSpawner(pub actix::ArbiterHandle);
-
-impl FutureSpawner for ActixArbiterHandleFutureSpawner {
-    fn spawn_boxed(&self, description: &'static str, f: BoxFuture<'static, ()>) {
-        if !self.0.spawn(f) {
-            near_o11y::tracing::error!(
-                "Failed to spawn future: {}, arbiter has exited",
-                description
-            );
-        }
     }
 }
 
@@ -138,29 +114,6 @@ impl<T> DelayedActionRunnerExt<T> for dyn DelayedActionRunner<T> + '_ {
         f: impl FnOnce(&mut T, &mut dyn DelayedActionRunner<T>) + Send + 'static,
     ) {
         self.run_later_boxed(name, dur, Box::new(f));
-    }
-}
-
-/// Implementation of `DelayedActionRunner` for Actix. With this, any code
-/// that used to take a `&mut actix::Context` can now take a
-/// `&mut dyn DelayedActionRunner<T>` instead, which isn't actix-specific.
-impl<T, Outer> DelayedActionRunner<T> for actix::Context<Outer>
-where
-    T: 'static,
-    Outer: DerefMut<Target = T>,
-    Outer: Actor<Context = actix::Context<Outer>>,
-{
-    fn run_later_boxed(
-        &mut self,
-        _name: &str,
-        dur: Duration,
-        f: Box<dyn FnOnce(&mut T, &mut dyn DelayedActionRunner<T>) + Send + 'static>,
-    ) {
-        near_performance_metrics::actix::run_later(
-            self,
-            dur.max(Duration::ZERO).unsigned_abs(),
-            move |obj, ctx| f(&mut *obj, ctx),
-        );
     }
 }
